Rename methods and variables

+++ core/src/main/java/in/testpress/util/TimeUtils.kt
@@ -2,7 +2,7 @@ package `in`.testpress.util
 
 object TimeUtils {
     @JvmStatic
-    fun convertMilliSecondsToSeconds(milliseconds: Long): Float {
-        return maxOf(0, milliseconds).toFloat() / 1000
+    fun convertMilliSecondsToSeconds(milliseconds: Long): Long {
+        return maxOf(0, milliseconds) / 1000
     }
 }
